% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-testing-plan.R
\name{auto_testing_text}
\alias{auto_testing_text}
\title{Generate automated testing boiler plate text for the testing plan}
\usage{
auto_testing_text(language = c("R", "Go"))
}
\arguments{
\item{language}{Denotes the language the package was coded in. Either 'R' or 'Go'.
Dictates boiler plate text in the generated validation docs.}
}
\description{
Generate automated testing boiler plate text for the testing plan
}
\keyword{internal}
