% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-test-framework.R
\name{check_input}
\alias{check_input}
\title{Check input for missing links between tests, requirements, and stories}
\usage{
check_input(dd)
}
\arguments{
\item{dd}{Tibble containing stories, requirements, and tests. Created in
\code{\link[=create_test_framework]{create_test_framework()}}.}
}
\description{
Check input for missing links between tests, requirements, and stories
}
