% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-checks.R
\name{check_inputs}
\alias{check_inputs}
\alias{find_tests_without_reqs}
\alias{find_reqs_with_missing_tests}
\alias{find_reqs_without_stories}
\title{Find missing input pieces.}
\usage{
check_inputs(merged_inputs)

find_tests_without_reqs(merged_inputs)

find_reqs_with_missing_tests(merged_inputs)

find_reqs_without_stories(merged_inputs)
}
\arguments{
\item{merged_inputs}{Tibble with stories, requirements and tests, as returned
by \code{\link[=create_validation_docs]{create_validation_docs()}}.}
}
\value{
Tibble with the missing items, or, in the case of \code{check_inputs()}, a
list of tibbles.
}
\description{
\code{find_tests_without_reqs()} returns IDs from the test results that are not
associated with a requirement. \code{find_reqs_with_missing_tests()} goes in the
reverse direction, returning requirements with test IDs that aren't found in
the test results. \code{find_reqs_without_stories()} returns requirements that are
not linked to a story.

\code{check_inputs()} is a convenience wrapper around all the above functions. It
returns their results in a list and also displays messages about the results.
}
\seealso{
\code{\link[=create_validation_docs]{create_validation_docs()}}, \link{input_formats}
}
