% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-checks.R
\name{check_test_input}
\alias{check_test_input}
\title{Check input test data.}
\usage{
check_test_input(tests)
}
\arguments{
\item{tests}{Tibble with TestId column}
}
\value{
Tibble with NA tests IDs removed.
}
\description{
\itemize{
\item Warn about rows without test IDs and filter them out.
\item Abort if non-NA IDs are not unique across rows.
}
}
\seealso{
\code{\link[=create_package_docs]{create_package_docs()}}, \code{\link[=create_metworx_docs]{create_metworx_docs()}}, \link{input_formats}
}
