% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cleanup_rmds}
\alias{cleanup_rmds}
\title{Delete copied RMD's used to render word documents}
\usage{
cleanup_rmds(
  output_dir,
  file_names = c(VAL_PLAN_FILE, TEST_PLAN_FILE, TEST_RESULTS_FILE, MAT_FILE, REQ_FILE,
    VAL_SUM_FILE, RLS_NOTES_FILE),
  append = NULL
)
}
\arguments{
\item{output_dir}{directory where word documents will be generated}

\item{file_names}{file names of every validation plan}

\item{append}{package or 'metworx' appended to file name}
}
\description{
Delete copied RMD's used to render word documents
}
\keyword{internal}
