% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-metworx-docs.R
\name{create_metworx_docs}
\alias{create_metworx_docs}
\title{Create validation docs for Metworx platform}
\usage{
create_metworx_docs(
  product_name,
  version,
  specs,
  release_notes_file = NULL,
  auto_test_dir = NULL,
  man_test_dir = NULL,
  roles = NULL,
  style_dir = NULL,
  output_dir = getwd(),
  write = TRUE,
  cleanup_rmd = TRUE
)
}
\arguments{
\item{product_name}{The product being validated.}

\item{version}{The version number of the product.}

\item{specs}{tibble of stories and requirements. See \link{input_formats}.}

\item{release_notes_file}{file path to a formatted markdown doc of release notes.
This should contain two top-level headers for "Changes and New Features" and
"Bug Fixes".}

\item{style_dir}{Directory that has style references for the generated docx
files. When generating each output file, Pandoc will be instructed to use
the reference file from this directory that has the same base name (if it
exists).}

\item{output_dir}{Directory to write the output documents to. Defaults to
working directory.}

\item{write}{Whether to create the output docs. Setting this to \code{FALSE} is
useful when you're just interested in the return value.}

\item{cleanup_rmd}{Whether to delete the copied RMD's after the word documents are generated.
Defaults to \code{TRUE}.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
As of \verb{mrgvalidate 2.0.0}, this function is a prototype for
creating Metworx platform docs with the new 7-document format. However, it has
been labeled "experimental" because it has not yet been fully vetted in
a real Metworx validation.
}
\seealso{
\code{\link[=create_package_docs]{create_package_docs()}}
}
