% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-metworx-docs.R, R/create-package-docs.R
\name{create_metworx_docs}
\alias{create_metworx_docs}
\alias{create_package_docs}
\title{Create validation docs}
\usage{
create_metworx_docs(
  product_name,
  version,
  specs,
  release_notes_file = NULL,
  auto_test_dir = NULL,
  man_test_dir = NULL,
  roles = NULL,
  style_dir = NULL,
  output_dir = getwd(),
  write = TRUE,
  cleanup_rmd = TRUE
)

create_package_docs(
  product_name,
  version,
  language = c("R", "Go"),
  repo_url,
  specs,
  release_notes_file,
  auto_test_dir = NULL,
  style_dir = NULL,
  output_dir = getwd(),
  write = TRUE,
  cleanup_rmd = TRUE
)
}
\arguments{
\item{product_name}{The product being validated.}

\item{version}{The version number of the product.}

\item{specs}{tibble of stories and requirements. See \link{input_formats}.}

\item{release_notes_file}{file path to a formatted markdown doc of release notes.}

\item{auto_test_dir, man_test_dir}{path to directories containing automatic
and manual test output files. See \link{input_formats}.}

\item{roles}{A data frame of user roles that, if specified, is inserted into
the requirements document.}

\item{style_dir}{Directory that has style references for the generated docx
files. When generating each output file, Pandoc will be instructed to use
the reference file from this directory that has the same base name (if it
exists).}

\item{output_dir}{Directory to write the output documents to. Defaults to
working directory.}

\item{write}{Whether to create the output docs. Setting this to \code{FALSE} is
useful when you're just interested in the return value.}

\item{cleanup_rmd}{Whether to delete the copied RMD's after the word documents are generated.
Defaults to \code{TRUE}.}

\item{language}{Denotes the language the package was coded in. Either 'R' or 'Go'.
Dictates boiler plate text in the generated validation docs.}

\item{repo_url}{Character string denoting the url of repository.}
}
\value{
In addition to creating the validation docs, a tibble that joins the
tests with \code{specs} is returned invisibly.
}
\description{
Create validation docs
}
\details{
For packages, release notes will come from the corresponding \code{NEWS.md} document.
}
\section{Functions}{
\itemize{
\item \code{create_metworx_docs()}: Create validation docs for metworx

\item \code{create_package_docs()}: Create validation docs for packages

}}
