% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-validation-docs.R
\name{create_test_framework}
\alias{create_test_framework}
\title{Format stories, tests, and other relevant information into usable objects}
\usage{
create_test_framework(
  product_name,
  specs,
  auto_test_dir = NULL,
  man_test_dir = NULL,
  type = "package"
)
}
\arguments{
\item{product_name}{The product being validated.}

\item{specs}{tibble of stories and requirements. See \link{input_formats}.}

\item{auto_test_dir, man_test_dir}{path to directories containing automatic
and manual test output files. See \link{input_formats}.}

\item{type}{character string denoting which type of docs you want to generate. Either "package" or "metworx"}
}
\description{
Format stories, tests, and other relevant information into usable objects
}
\keyword{internal}
