% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-docs.R
\name{create_validation_docs}
\alias{create_validation_docs}
\title{Create validation docs}
\usage{
create_validation_docs(requirements, test_output_dir, output_dir = getwd())
}
\arguments{
\item{requirements}{tibble of requirements in the format returned by
\code{\link[=read_requirements_github]{read_requirements_github()}} or \code{\link[=read_requirements_gsheet]{read_requirements_gsheet()}}. It must have
the following columns: title, story, risk, and test_ids.}

\item{test_output_dir}{path to direction containing a test output files. See
\link{input-formats}.}

\item{output_dir}{Directory to write the output documents to. Defaults to
working directory.}
}
\description{
This function is the main entry point for creating validation docs.
}
