% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-templates.R
\name{create_validation_templates}
\alias{create_validation_templates}
\title{Create template docs for packages or metworx}
\usage{
create_validation_templates(
  type = c("package", "metworx"),
  output_dir = system.file(package = "mrgvalidate"),
  new_folder = "template_renders",
  style_dir = NULL,
  cleanup_rmd = TRUE
)
}
\arguments{
\item{type}{the type of doc you want to render ("package" or "metworx")}

\item{output_dir}{Directory to write the output documents to. Defaults to \code{system.file(package = "mrgvalidate")}.}

\item{new_folder}{Folder within \code{output_dir} to render docs.
Mainly useful when generating docs to a system directory that doesn't already exist}

\item{style_dir}{Directory to check for a docx style reference}

\item{cleanup_rmd}{Whether to delete the copied RMD's after the word documents are generated.
Defaults to \code{TRUE}.}
}
\description{
Create template docs for packages or metworx
}
\details{
\subsection{Developer Notes:}{
\itemize{
\item \strong{Note:} This function does \strong{not} call the internal \code{make_} functions responsible for creating the validation
docs under normal circumstances.
\item \code{system.file()} wont register nested directories in the file path if the folder doesn't exist. In other words,
\code{output_dir = system.file("template_renders", package = "mrgvalidate")} would return \code{""}, and the function would fail
(because \code{"template_renders"} does not exist within \code{inst}). Use the \code{new_folder} argument when rendering to a
new folder within \code{inst}.
}
}
}
