% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_bug_section}
\alias{extract_bug_section}
\title{Extract bugs section from release notes character vector}
\usage{
extract_bug_section(notes_lines)
}
\arguments{
\item{notes_lines}{release notes character vector}
}
\description{
Extract bugs section from release notes character vector
}
\keyword{internal}
