% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-checks.R
\name{find_tests_without_reqs}
\alias{find_tests_without_reqs}
\alias{find_reqs_with_missing_tests}
\alias{find_reqs_without_stories}
\title{Find missing input pieces.}
\usage{
find_tests_without_reqs(merged_inputs)

find_reqs_with_missing_tests(merged_inputs)

find_reqs_without_stories(merged_inputs)
}
\arguments{
\item{merged_inputs}{Tibble with stories, requirements and tests, as returned
by \code{\link[=create_validation_docs]{create_validation_docs()}}.}
}
\value{
Tibble with the missing items.
}
\description{
\code{find_tests_without_reqs()} returns IDs from the test results that are not
associated with a requirement. \code{find_reqs_with_missing_tests()} goes in the
reverse direction, returning requirements with test IDs that aren't found in
the test results. \code{find_reqs_without_stories()} returns requirements that are
not linked to a story.
}
\seealso{
\code{\link[=create_validation_docs]{create_validation_docs()}}, \link{input_formats}
}
