% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-validation-summary-report.R
\name{format_bugs}
\alias{format_bugs}
\title{format bugs from release_notes}
\usage{
format_bugs(release_notes)
}
\arguments{
\item{release_notes}{a list of release_notes}
}
\description{
format bugs from release_notes
}
\keyword{internal}
