% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-validation-plan.R
\name{format_release_changes}
\alias{format_release_changes}
\title{format function changes from release_notes}
\usage{
format_release_changes(release_notes = NULL)
}
\arguments{
\item{release_notes}{a list of release_notes}
}
\description{
format function changes from release_notes
}
\keyword{internal}
