% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_req_tests}
\alias{format_req_tests}
\title{format tests for inclusion in requirements specification document}
\usage{
format_req_tests(x)
}
\arguments{
\item{x}{A single row from the stories df in \code{\link[=make_requirements]{make_requirements()}}}
}
\description{
format tests for inclusion in requirements specification document
}
\keyword{internal}
