% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_rmd_name}
\alias{format_rmd_name}
\title{Rename template}
\usage{
format_rmd_name(output_dir, out_file, append = NULL)
}
\arguments{
\item{output_dir}{directory for \code{out_file} to be copied to}

\item{out_file}{name of file, including extension. Note: this is not a file path.}

\item{append}{character string to append to out_file. Will be separated by '-'.}
}
\description{
Rename template
}
\keyword{internal}
