% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-requirements.R
\name{format_roles}
\alias{format_roles}
\title{Format roles in requirements specification}
\usage{
format_roles(roles)
}
\arguments{
\item{roles}{A data frame of user roles. If specified, this will be
inserted as a table under a "User Roles" section.}
}
\description{
Format roles in requirements specification
}
\keyword{internal}
