% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_spec}
\alias{format_spec}
\title{format a story for inclusion in output documents}
\usage{
format_spec(x)
}
\arguments{
\item{x}{A single row from the stories df in \code{\link[=make_requirements]{make_requirements()}}}
}
\description{
format a story for inclusion in output documents
}
\keyword{internal}
