% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-testing-results.R
\name{format_val_info}
\alias{format_val_info}
\title{Format info on automatic tests}
\usage{
format_val_info(auto_info)
}
\arguments{
\item{auto_info}{named list}
}
\description{
Format info on automatic tests
}
\keyword{internal}
