% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-docs.R
\name{generate_docs}
\alias{generate_docs}
\title{Wrapper to generate all three documents, using defaults for output paths}
\usage{
generate_docs(
  org,
  repo,
  milestone,
  version,
  domain = VALID_DOMAINS,
  extra_test_dirs = NULL,
  output_dir = getwd()
)
}
\arguments{
\item{org}{Github organization that the repo is under}

\item{repo}{The name of the repo for the package you are validating}

\item{milestone}{The name of the milestone associated with the release you are validating. All issues tied to this milestone with be pulled.}

\item{version}{The version number of the package you are validating. This must correspond to a tag in the repo.}

\item{domain}{Domain where repo lives. Either "github.com" or "ghe.metrumrg.com", defaulting to "github.com"}

\item{extra_test_dirs}{Character vector of paths (relative to package root dir) to directories that contain additional tests to run}

\item{output_dir}{Directory to write the output documents to. Defaults to working directory.}
}
\description{
Wrapper to generate all three documents, using defaults for output paths
}
