% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_reference_docx}
\alias{get_reference_docx}
\title{Return reference docx file whose name matches current output file.}
\usage{
get_reference_docx(out_file, style_dir)
}
\arguments{
\item{out_file, style_dir}{The output file and reference directory passed to a
\verb{write_*} function (e.g., \code{\link[=make_requirements]{make_requirements()}}).}
}
\value{
The full path of the reference docx within \code{style_dir} with the same
name as \code{out_file}; if such a file doesn't exist, return the default value
\code{reference_docx} value of \code{\link[rmarkdown:word_document]{rmarkdown::word_document()}}.
}
\description{
Return reference docx file whose name matches current output file.
}
\keyword{internal}
