% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-validation-docs.R
\name{get_repo_vars}
\alias{get_repo_vars}
\title{Get repo URL and commit hash}
\usage{
get_repo_vars(auto_info, type)
}
\arguments{
\item{auto_info}{named list}

\item{type}{either "package" or "metworx"}
}
\description{
Get repo URL and commit hash
}
\details{
Environment variables not set, but included in \code{mrgvalprep::get_sys_info}, will show up as an empty character string ('')
If these variables are not set or included in \code{mrgvalprep::get_sys_info}, values will show up as NULL.
This function accounts for both cases, and will only fetch these values if not specified.
This method was chosen to preserve existing functionality of validation build scripts.
}
\keyword{internal}
