% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-sha.R
\name{get_sha}
\alias{get_sha}
\title{Get the SHA matching a git reference}
\usage{
get_sha(repo, ref, type = c("tag", "branch"))
}
\arguments{
\item{repo}{repository address in the format \code{username/repo}}

\item{ref}{name of either a tag or a branch}

\item{type}{type of reference for \code{ref}}
}
\value{
the SHA for \code{ref}
}
\description{
The \code{remotes} package provides nice functionality for handling a variety of
reference types, but many of the necessary functions are not exported.
Because we only need to consider two types, copying all that machinery is
probably more trouble than maintaining our own lightweight implementation.
}
\keyword{internal}
