% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-sha.R
\name{get_sha_from_tags}
\alias{get_sha_from_tags}
\title{Get the commit SHA matching a tag name}
\usage{
get_sha_from_tags(tags, name)
}
\arguments{
\item{tags}{a list of tags, as returned by the \code{tags} API endpoint}

\item{name}{the tag name}
}
\value{
the SHA of the commit associated with tag \code{name}
}
\description{
There is no endpoint to get the commit SHA directly from a tag. Instead, we
can return the matching element from the set of tags.
}
\keyword{internal}
