% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_template}
\alias{get_template}
\title{fetch template from package}
\usage{
get_template(
  template = c("validation_plan", "testing_plan", "testing_results",
    "traceability_matrix", "requirements_specification", "validation_summary_report"),
  type = c("package", "metworx")
)
}
\arguments{
\item{template}{string matching which template you want to render}

\item{type}{the type of doc you want to render ("package" or "metworx")}
}
\description{
fetch template from package
}
