% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-checks.R
\name{has_req_cols}
\alias{has_req_cols}
\title{Return TRUE if \code{merged_inputs} has requirement columns.}
\usage{
has_req_cols(merged_inputs)
}
\description{
Return TRUE if \code{merged_inputs} has requirement columns.
}
\keyword{internal}
