% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_formats.R
\name{input-formats}
\alias{input-formats}
\title{mrgvalidate test result input}
\description{
In order to format validation docs, mrgvalidate must collect various pieces
of information from the caller. This information can be broken down into two
categories: information about the requirements and information about the
executed tests. Below is a description of how information about the executed
tests is expected to be formatted.
}
\details{
For automated tests, a directory should contain set of CSVs with the test
results. A CSV file must have the following columns:
\itemize{
\item test_name: a description of the test
\item passed, failed: a count of passed and failed assertions for the test
\item test_tag: a unique identifier of the test. This field is used to map tests
to requirements.
}

For \code{testthat} tests, the \code{\link[=parse_testthat_list_reporter]{parse_testthat_list_reporter()}} helper can be used
to convert \link[testthat:ListReporter]{testthat::ListReporter} results into the above format.

Alongside each CSV, there must be a JSON file with the same base name. This
file includes information about the run itself (e.g., the date it was
executed and system details). At the minimum, it must have a "date" field.
Any key-value pairs that are specified under the top-level "info" field will
be rendered in the validation docs.\preformatted{\{
    "date": "2021-07-26 12:20:10 EDT",
    "info":
    \{
        "commit": "78f70b9297a7f8b0f2ec2a18d17a9cc6722359c8"
    \}
\}
}
}
