% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_temp_github_package}
\alias{install_temp_github_package}
\title{Install a GitHub package to a temporary location}
\usage{
install_temp_github_package(repo, ref)
}
\arguments{
\item{repo}{Repository address in the format
\verb{username/repo[/subdir][@ref|#pull]}. Alternatively, you can
specify \code{subdir} and/or \code{ref} using the respective parameters
(see below); if both is specified, the values in \code{repo} take
precedence.}

\item{ref}{Desired git reference. Could be a commit, tag, or branch
name, or a call to \code{\link[remotes:github_pull]{github_pull()}}. Defaults to \code{"master"}.}
}
\value{
the temporary directory, invisibly
}
\description{
Install a GitHub package to a temporary location
}
\keyword{internal}
