% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_with_tests}
\alias{install_with_tests}
\title{Install a GitHub package with tests}
\usage{
install_with_tests(repo, ref = "HEAD", ...)
}
\arguments{
\item{repo}{Repository address in the format
\verb{username/repo[/subdir][@ref|#pull]}. Alternatively, you can
specify \code{subdir} and/or \code{ref} using the respective parameters
(see below); if both is specified, the values in \code{repo} take
precedence.}

\item{ref}{Desired git reference. Could be a commit, tag, or branch
name, or a call to \code{\link[remotes:github_pull]{github_pull()}}. Defaults to \code{"master"}.}

\item{...}{additional arguments passed to \code{\link[remotes:install_github]{remotes::install_github()}}}
}
\description{
This is a light wrapper around \code{\link[remotes:install_github]{remotes::install_github()}}.
}
\keyword{internal}
