% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-release-notes.R
\name{make_release_notes}
\alias{make_release_notes}
\title{Build the Release Notes document and write it to output file(s)}
\usage{
make_release_notes(
  product_name,
  version,
  repo_url = NULL,
  release_notes = NULL,
  style_dir = NULL,
  out_file = RLS_NOTES_FILE,
  output_dir = getwd(),
  type = "package",
  word_document = TRUE
)
}
\arguments{
\item{product_name}{The name of the product you are validating, to be included in the output document.}

\item{version}{The version number of the product you are validating, to be included in the output document.}

\item{repo_url}{Character string denoting the url of repository.}

\item{release_notes}{list of release notes, formatted for rmarkdown.}

\item{style_dir}{Directory to check for a docx style reference that has the
same base name as \code{out_file}.}

\item{out_file}{Filename to write markdown file out to. Any extension will be ignored and replaced with .Rmd}

\item{output_dir}{Directory to write the output documents to. Defaults to working directory.}

\item{type}{the type of doc you want to render ("package" or "metworx")}

\item{word_document}{Logical scaler indicating whether to render a docx document}
}
\description{
Build the Release Notes document and write it to output file(s)
}
\keyword{internal}
