% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_signature_line}
\alias{make_signature_line}
\title{Make signature page}
\usage{
make_signature_line()
}
\description{
Make signature page
}
\details{
The template word docs were modified so that Heading 9 has a bottom border, which functions as a custom signature line
}
\keyword{internal}
