% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-traceability-matrix.R
\name{make_traceability_matrix}
\alias{make_traceability_matrix}
\title{Build the Traceability Matrix and write it to output file(s)}
\usage{
make_traceability_matrix(
  product_name,
  version,
  df,
  style_dir = NULL,
  out_file = MAT_FILE,
  output_dir = getwd(),
  type = "package",
  word_document = TRUE
)
}
\arguments{
\item{product_name}{The name of the product you are validating, to be included in the output document.}

\item{version}{The version number of the product you are validating, to be included in the output document.}

\item{df}{Tibble containing stories, requirements, and tests. Created in
\code{\link[=create_test_framework]{create_test_framework()}}.}

\item{style_dir}{Directory to check for a docx style reference that has the
same base name as \code{out_file}.}

\item{out_file}{filename to write markdown file out to. Any extension will be ignored and replaced with .md}

\item{output_dir}{Directory to write the output documents to. Defaults to working directory.}

\item{type}{the type of doc you want to render ("package" or "metworx")}

\item{word_document}{Logical scaler indicating whether to render a docx document}
}
\description{
Build the Traceability Matrix and write it to output file(s)
}
\keyword{internal}
