% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgvalidate.R
\name{mrgvalidate}
\alias{mrgvalidate}
\title{Create validation documents from Stories, Requirements, and test outputs}
\description{
The \code{mrgvalidate} package ingests structured data including text describing
user stories and technical requirements, and testing outputs, and creates
validation documents by joining these together and inserting them into
pre-defined markdown templates that can then be automatically rendered into
\code{.docx} format. Specifications for input data can be found on the
\link{input_formats} help page. The companion \code{mrgvalprep} package contains
helpers for formatting input data to the specification in \link{input_formats}.
}
