% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-validation-testing.R
\name{print_info_list}
\alias{print_info_list}
\title{Format an info JSON suite element for printing}
\usage{
print_info_list(.l)
}
\description{
Format an info JSON suite element for printing
}
\keyword{internal}
