% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-results.R
\name{read_csv_test_results}
\alias{read_csv_test_results}
\title{Read test results from CSV files in a directory.}
\usage{
read_csv_test_results(test_output_dir)
}
\arguments{
\item{test_output_dir}{path to a directory containing test output files. Each
CSV file should have an accompanying JSON file. See \link{input_formats}.}
}
\value{
List of two elements, a tibble with test result ("results") and
metadata information for each test result set ("info").
}
\description{
Read test results from CSV files in a directory.
}
