% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-results.R
\name{read_manual_test_results}
\alias{read_manual_test_results}
\title{Read manual test results.}
\usage{
read_manual_test_results(test_output_dir)
}
\arguments{
\item{test_output_dir}{Directory that contains test subdirectories named by
test ID.}
}
\value{
Tibble with TestId and content columns. In the content, links to the
test's assets subdirectory are switched from relative to absolute paths.
}
\description{
Read manual test results.
}
