% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-requirements.R
\name{read_requirements_gsheet}
\alias{read_requirements_gsheet}
\title{Read requirements from a Google Sheet.}
\usage{
read_requirements_gsheet(
  ss,
  sheet = NULL,
  title_col = "title",
  story_col = "story",
  test_ids_col = "test_id",
  risk_col = "risk"
)
}
\arguments{
\item{ss, sheet}{Sheet identifiers passed \code{\link[googlesheets4:range_read]{googlesheets4::read_sheet()}}.}

\item{title_col, story_col, risk, test_ids_col}{Names to remap to title, story,
risk, and test_ids columns.}
}
\value{
Tibble with these columns: title, story, risk, and test_ids
}
\description{
Read requirements from a Google Sheet.
}
