% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-installed-tests.R
\name{run_installed_tests}
\alias{run_installed_tests}
\title{Run tests for an installed package}
\usage{
run_installed_tests(
  package,
  path = NULL,
  reporter = testthat::ListReporter$new(),
  unload = FALSE
)
}
\arguments{
\item{package}{Name of installed package.}

\item{path}{Path to directory containing tests.}

\item{reporter}{Reporter to use to summarise output. Can be supplied
as a string (e.g. "summary") or as an R6 object
(e.g. \code{SummaryReporter$new()}).

See \link[testthat]{Reporter} for more details and a list of built-in reporters.}

\item{unload}{\code{TRUE} to unload \code{pkg} after running its tests}
}
\value{
list
}
\description{
\code{package} must have been installed with \code{--install-tests}
}
