% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_results.R
\name{test_results}
\alias{test_results}
\title{Get test results for a package}
\usage{
test_results(dry_run = TRUE, repo = NULL, ref = NULL, ...)
}
\arguments{
\item{dry_run}{\code{TRUE} to look for test results in \code{all_tests.csv}, \code{FALSE} to
download the package and run the tests}

\item{repo}{Repository address in the format
\verb{username/repo[/subdir][@ref|#pull]}. Alternatively, you can
specify \code{subdir} and/or \code{ref} using the respective parameters
(see below); if both is specified, the values in \code{repo} take
precedence.}

\item{ref}{Desired git reference. Could be a commit, tag, or branch
name, or a call to \code{\link[remotes:github_pull]{github_pull()}}. Defaults to \code{"master"}.}

\item{...}{additional arguments passed to \code{\link[=validate_tests]{validate_tests()}}}
}
\value{
the test results, as a tibble
}
\description{
Get test results for a package
}
