% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-traceability-matrix.R
\name{write_traceability_matrix}
\alias{write_traceability_matrix}
\title{Build the Traceability Matrix and write it to output file(s)}
\usage{
write_traceability_matrix(
  df,
  product,
  version,
  out_file = MAT_FILE,
  output_dir = getwd(),
  word_document = TRUE
)
}
\arguments{
\item{df}{Tibble containing stories, requirements, and tests. Created in
\code{\link[=create_validation_docs]{create_validation_docs()}}.}

\item{product}{The name of the product you are validating, to be included in the output document.}

\item{version}{The version number of the product you are validating, to be included in the output document.}

\item{out_file}{filename to write markdown file out to. Any extension will be ignored and replaced with .md}

\item{output_dir}{Directory to write the output documents to. Defaults to working directory.}

\item{word_document}{Logical scaler indicating whether to render a docx document}
}
\description{
Build the Traceability Matrix and write it to output file(s)
}
\keyword{internal}
