% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-validation-testing.R
\name{write_validation_testing}
\alias{write_validation_testing}
\title{Build the Validation Testing document and write it to a markdown file}
\usage{
write_validation_testing(
  product,
  version,
  tests,
  auto_info,
  style_dir = NULL,
  out_file = VAL_FILE,
  output_dir = getwd(),
  word_document = TRUE
)
}
\arguments{
\item{product}{The name of the product you are validating, to be included in the output document.}

\item{version}{The version number of the product you are validating, to be included in the output document.}

\item{tests}{Tibble containing all test results, FORMAT: CREATED ON LINE 59
OF \code{generate-docs.R} in \code{\link[=create_validation_docs]{create_validation_docs()}}. ** TODO: If we're gonna
export this function we're gonna need to spell out what this tibble
actually needs to look like so that someone could theoretically constuct
it. We may just wanna make these \verb{write_*} functions private though. When
would someone need to call them?**}

\item{auto_info}{A named list containing the test suite information pulled
from the \code{.json} files found in \code{auto_test_dir}, one element per \code{.json}
(named with the filename \emph{without} extension). \strong{Same note as \code{tests} about
exporting and specs.}}

\item{style_dir}{Directory to check for a docx style reference that has the
same base name as \code{out_file}.}

\item{out_file}{Filename to write markdown file out to. Any extension will be ignored and replaced with .md}

\item{output_dir}{Directory to write the output documents to. Defaults to working directory.}

\item{word_document}{Logical scaler indicating whether to render a docx document}
}
\description{
This takes the input from automated and/or manual tests and writes them to a
\code{.md} file (and optionally \code{.docx} file as well). The automated tests will be
written as tables of results, one per result file (\code{.csv} file found in
\code{auto_test_dir} passed to \code{\link[=create_validation_docs]{create_validation_docs()}}). The manual tests will
have their \code{content} sections rendered sequentially below the automated
tests.
}
