% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-test-framework.R
\name{create_test_framework}
\alias{create_test_framework}
\title{Format stories, tests, and other relevant information into usable objects}
\usage{
create_test_framework(
  product_name,
  specs,
  auto_test_dir = NULL,
  man_test_dir = NULL,
  type = "package"
)
}
\arguments{
\item{product_name}{The product being validated.}

\item{specs}{tibble of stories and requirements. See \link{input_formats}.}
}
\description{
Format stories, tests, and other relevant information into usable objects
}
\keyword{internal}
