% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-docs.R
\name{create_validation_docs}
\alias{create_validation_docs}
\title{Create validation docs}
\usage{
create_validation_docs(requirements, test_output_dir, output_dir = getwd())
}
\arguments{
\item{test_output_dir}{path to direction containing a test output files. See
\link{input_formats}.}

\item{output_dir}{Directory to write the output documents to. Defaults to
working directory.}

\item{specs}{tibble of requirements in the format returned by
\code{\link[=read_spec_gsheets]{read_spec_gsheets()}}.}
}
\description{
This function is the main entry point for creating validation docs.
}
