% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-docs.R
\name{create_validation_docs}
\alias{create_validation_docs}
\title{Create validation docs}
\usage{
create_validation_docs(
  product_name,
  version,
  specs,
  auto_test_dir = NULL,
  man_test_dir = NULL,
  roles = NULL,
  style_dir = NULL,
  output_dir = getwd(),
  write = TRUE
)
}
\arguments{
\item{product_name}{The product being validated.}

\item{version}{The version number of the product.}

\item{specs}{tibble of requirements in the format returned by
\code{\link[=read_spec_gsheets]{read_spec_gsheets()}}.}

\item{auto_test_dir, man_test_dir}{path to directories containing automatic
and manual test output files. See \link{input_formats}.}

\item{roles}{A data frame of user roles that, if specified, is inserted into
the requirements document.}

\item{style_dir}{Directory that has style references for the generated docx
files. When generating each output file, Pandoc will be instructed to use
the reference file from this directory that has the same base name (if it
exists).}

\item{output_dir}{Directory to write the output documents to. Defaults to
working directory.}

\item{write}{Whether to create the output docs. Setting this to \code{FALSE} is
useful when you're just interested in the return value.}
}
\value{
In addition to creating the validation docs, a tibble that joins the
tests with \code{specs} is returned invisibly.
}
\description{
This function is the main entry point for creating validation docs.
}
