% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-metworx-docs.R, R/create-package-docs.R,
%   R/create-validation-docs.R
\name{create_metworx_docs}
\alias{create_metworx_docs}
\alias{create_package_docs}
\alias{create_validation_docs}
\title{Create validation docs}
\usage{
create_metworx_docs(
  product_name,
  version,
  specs,
  release_notes_file = NULL,
  auto_test_dir = NULL,
  man_test_dir = NULL,
  roles = NULL,
  style_dir = NULL,
  output_dir = getwd(),
  write = TRUE,
  cleanup_rmd = TRUE
)

create_package_docs(
  product_name,
  version,
  repo_url,
  specs,
  release_notes_file,
  auto_test_dir = NULL,
  man_test_dir = NULL,
  style_dir = NULL,
  output_dir = getwd(),
  write = TRUE,
  cleanup_rmd = TRUE
)

create_validation_docs(
  product_name,
  version,
  specs,
  auto_test_dir = NULL,
  man_test_dir = NULL,
  release_notes_file = NULL,
  roles = NULL,
  style_dir = NULL,
  output_dir = getwd(),
  type = c("package", "metworx"),
  write = TRUE
)
}
\arguments{
\item{product_name}{The product being validated.}

\item{version}{The version number of the product.}

\item{specs}{tibble of stories and requirements. See \link{input_formats}.}

\item{release_notes_file}{file path to a formatted markdown doc of release notes.}

\item{auto_test_dir, man_test_dir}{path to directories containing automatic
and manual test output files. See \link{input_formats}.}

\item{roles}{A data frame of user roles that, if specified, is inserted into
the requirements document.}

\item{style_dir}{Directory that has style references for the generated docx
files. When generating each output file, Pandoc will be instructed to use
the reference file from this directory that has the same base name (if it
exists).}

\item{output_dir}{Directory to write the output documents to. Defaults to
working directory.}

\item{write}{Whether to create the output docs. Setting this to \code{FALSE} is
useful when you're just interested in the return value.}

\item{cleanup_rmd}{Whether to delete the copied RMD's after the word documents are generated.
Defaults to \code{TRUE}.}

\item{repo_url}{Character string denoting the url of repository.}

\item{type}{character string denoting which type of docs you want to generate. Either "package" or "metworx"}
}
\value{
In addition to creating the validation docs, a tibble that joins the
tests with \code{specs} is returned invisibly.

In addition to creating the validation docs, a tibble that joins the
tests with \code{specs} is returned invisibly.
}
\description{
This function is the main entry point for creating validation docs.
}
\details{
For packages, release notes will come from the corresponding \code{NEWS.md} document.
This should be copied into \code{auto_test_dir} in the build script before creating the docs via \code{create_validation_docs} or \code{create_package_docs}:
\code{fs::file_copy("NEWS.md", file.path(auto_test_dir, "NEWS.md"))}

For metworx, a formatted markdown file path must be provided via the \code{release_notes_file} argument.

Make sure sections are separated by \verb{Heading 2 (##)} in both cases.
}
\section{Functions}{
\itemize{
\item \code{create_metworx_docs()}: Create validation docs for metworx

\item \code{create_package_docs()}: Create validation docs for packages

}}
