% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-validation-docs.R
\name{create_validation_docs}
\alias{create_validation_docs}
\title{Create validation docs}
\usage{
create_validation_docs(
  product_name,
  version,
  specs,
  auto_test_dir = NULL,
  man_test_dir = NULL,
  roles = NULL,
  output_dir = getwd()
)
}
\arguments{
\item{product_name}{The product being validated.}

\item{version}{The version number of the product.}

\item{specs}{tibble of stories and requirements. See \link{input_formats}.}

\item{auto_test_dir, man_test_dir}{path to directories containing automatic
and manual test output files. See \link{input_formats}.}

\item{roles}{A data frame of user roles that, if specified, is inserted into
the requirements document.}

\item{output_dir}{Directory to write the output documents to. Defaults to
working directory.}
}
\description{
This function is the main entry point for creating validation docs.
}
