% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{flextable_word}
\alias{flextable_word}
\title{Autofit flextables in word}
\usage{
flextable_word(tab, pg_width = 7, column_shrink = NULL)
}
\arguments{
\item{tab}{a flextable object}

\item{pg_width}{width (in inches) of word document}

\item{column_shrink}{character string. If specified, shrink this column before fitting to word document}
}
\description{
\code{flextable} by default will make the tables as wide as possible in word. This function will correct the \code{autofit()} feature and make the contents fit.
}
\keyword{internal}
