% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-testing-results.R
\name{format_auto_tests_results}
\alias{format_auto_tests_results}
\title{Format automatic test plan into flextable in for word doc rendering}
\usage{
format_auto_tests_results(auto_tests, val_info)
}
\arguments{
\item{auto_tests}{dataframe of automatic tests}
}
\description{
Format automatic test plan into flextable in for word doc rendering
}
\keyword{internal}
