% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-testing-plan.R
\name{format_man_test_plan}
\alias{format_man_test_plan}
\title{Format manual test plan into flextable in for word doc rendering}
\usage{
format_man_test_plan(man_tests)
}
\arguments{
\item{test_df}{dataframe of manual tests}
}
\description{
Format manual test plan into flextable in for word doc rendering
}
\details{
this also includes text, as we do not want to render this section in the absence of manual tests
}
\keyword{internal}
