% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_release_notes}
\alias{parse_release_notes}
\title{Parse release notes and filter to current version}
\usage{
parse_release_notes(release_notes, product_name, version, section_break = "##")
}
\arguments{
\item{release_notes}{lines of NEWS.md}

\item{product_name}{name of package or metworx blueprint}

\item{version}{the product's version number (must match NEWS.md)}
}
\description{
Parse release notes and filter to current version
}
\keyword{internal}
