% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-requirements.R
\name{read_requirements_gsheet}
\alias{read_requirements_gsheet}
\title{Read requirements from a Google Sheet.}
\usage{
read_requirements_gsheet(
  ss,
  sheet = NULL,
  req_id_col = "RequirementID",
  req_description_col = "Requirement Description",
  test_ids_col = "TestIDs"
)
}
\arguments{
\item{ss, sheet}{Sheet identifiers passed \code{\link[googlesheets4:range_read]{googlesheets4::read_sheet()}}.}

\item{req_id_col, req_description_col, test_ids_col}{Names to remap to
requirement ID, description, and test IDs.}
}
\value{
Tibble with these columns: req_id, req_description, and test_ids
}
\description{
Read requirements from a Google Sheet.
}
