% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-requirements.R
\name{write_requirements}
\alias{write_requirements}
\title{Build the Requirements Specification and write it to output file(s)}
\usage{
write_requirements(
  df,
  product_name,
  version,
  roles = NULL,
  style_dir = NULL,
  out_file = REQ_FILE,
  output_dir = getwd(),
  word_document = TRUE
)
}
\arguments{
\item{df}{Tibble containing stories, requirements, and tests. Created in
\code{\link[=create_validation_docs]{create_validation_docs()}}.}

\item{product_name}{The product you are validating, to be included in the output document.}

\item{version}{The version number of the product you are validating, to be included in the output document.}

\item{roles}{A data frame of user roles. If specified, this will be
inserted as a table under a "User Roles" section.}

\item{style_dir}{Directory to check for a docx style reference that has the
same base name as \code{out_file}.}

\item{out_file}{filename to write markdown file out to. Any extension will be ignored and replaced with .md}

\item{output_dir}{Directory to write the output documents to. Defaults to working directory.}

\item{word_document}{Logical scaler indicating whether to render a docx document}
}
\description{
Build the Requirements Specification and write it to output file(s)
}
\keyword{internal}
